+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.160.0](https://github.com/lodgify/lodgify-ui/compare/v1.159.0...v1.160.0) (2019-03-26)
+
+
+### Features
+
+* **react-dates-datepicker:** add styles for blocked and unblocked dates ([3b09984](https://github.com/lodgify/lodgify-ui/commit/3b09984))
+
+# [1.159.0](https://github.com/lodgify/lodgify-ui/compare/v1.158.0...v1.159.0) (2019-03-26)
+
+
+### Features
+
+* **SearchBar:** expose props to control input values ([1b3172b](https://github.com/lodgify/lodgify-ui/commit/1b3172b))
+
 # [1.158.0](https://github.com/lodgify/lodgify-ui/compare/v1.157.0...v1.158.0) (2019-03-21)